+++ features/playground-controller.js
@@ -1,18 +1,18 @@
-import { ThemeManager } from '../modules/theme-manager.js?v=20260630-3';
-import { FontManager } from '../modules/font-manager.js?v=20260630-3';
-import { UIManager } from '../modules/ui-manager.js?v=20260630-3';
-import { PanelManager } from '../modules/panel-manager.js?v=20260630-3';
-import { FileManager } from '../modules/file-manager.js?v=20260630-3';
-import { GraphManager } from '../modules/graph-manager.js?v=20260630-3';
-import { PopupManager } from '../modules/popup-manager.js?v=20260630-3';
-import { ClingoManager } from '../modules/clingo-manager.js?v=20260630-3';
-import { OutputManager } from '../modules/output-manager.js?v=20260630-3';
-import { ExportManager } from '../modules/export-manager.js?v=20260630-3';
-import { MetricsManager } from '../modules/metrics-manager.js?v=20260630-3';
-import { ConfigController } from './config-controller.js?v=20260630-3';
-import { DocsController } from './docs-controller.js?v=20260630-3';
-import { EditorController } from './editor-controller.js?v=20260630-3';
-import { ExamplesController } from './examples-controller.js?v=20260630-3';
+import { ThemeManager } from '../modules/theme-manager.js?v=20260630-4';
+import { FontManager } from '../modules/font-manager.js?v=20260630-4';
+import { UIManager } from '../modules/ui-manager.js?v=20260630-4';
+import { PanelManager } from '../modules/panel-manager.js?v=20260630-4';
+import { FileManager } from '../modules/file-manager.js?v=20260630-4';
+import { GraphManager } from '../modules/graph-manager.js?v=20260630-4';
+import { PopupManager } from '../modules/popup-manager.js?v=20260630-4';
+import { ClingoManager } from '../modules/clingo-manager.js?v=20260630-4';
+import { OutputManager } from '../modules/output-manager.js?v=20260630-4';
+import { ExportManager } from '../modules/export-manager.js?v=20260630-4';
+import { MetricsManager } from '../modules/metrics-manager.js?v=20260630-4';
+import { ConfigController } from './config-controller.js?v=20260630-4';
+import { DocsController } from './docs-controller.js?v=20260630-4';
+import { EditorController } from './editor-controller.js?v=20260630-4';
+import { ExamplesController } from './examples-controller.js?v=20260630-4';
 
 export class PlaygroundController {
     constructor(dom, store) {
@@ -93,6 +93,8 @@ export class PlaygroundController {
         );
 
         this.docsController.init();
+        const initialGraphMode = this.dom.graphModeRadios.find((radio) => radio.checked)?.value || 'standard';
+        this.docsController.setLegendMode(initialGraphMode);
         this.attachEventListeners();
 
         UIManager.initializeEmptyStates();
@@ -144,6 +146,7 @@ export class PlaygroundController {
 
         this.dom.graphModeRadios.forEach((radio) => {
             radio.addEventListener('change', () => {
+                this.docsController.setLegendMode(radio.value);
                 this.outputManager.clearActiveExtension();
                 this.graphManager.resetGraphColors();
                 this.pendingGraphUpdate = this.regenerateGraph();